Feat: 읽고 있는 책 목록 조회 API 추가

+++ ilkda/src/main/java/com/ilkda/server/record/model/Record.java
@@ -26,7 +26,7 @@ public class Record extends BaseEntity {
 
     private String report;
 
-    private int readPage;
+    private Long readPage;
 
     private boolean complete;
 
@@ -36,7 +36,7 @@ public class Record extends BaseEntity {
         this.member = member;
         this.book = book;
         this.report = "";
-        this.readPage = 0;
+        this.readPage = 0L;
         this.complete = false;
     }
 }
